refactor: menu, product 패키지 결합
