+++ src/main/java/kitchenpos/menu/application/dto/response/ProductResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.product.application.dto.response;
+package kitchenpos.menu.application.dto.response;
 
 import java.math.BigDecimal;
 import java.util.Objects;
