+++ src/test/java/kitchenpos/integration/MenuIntegrationTest.java
@@ -12,7 +12,7 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import kitchenpos.integration.helper.InitIntegrationTest;
 import kitchenpos.menu.application.dto.response.MenuGroupResponse;
 import kitchenpos.menu.application.dto.response.MenuResponse;
-import kitchenpos.product.application.dto.response.ProductResponse;
+import kitchenpos.menu.application.dto.response.ProductResponse;
 import org.junit.jupiter.api.Test;
 import java.math.BigDecimal;
 import java.util.List;
