+++ src/test/java/kitchenpos/integration/TableIntegrationTest.java
@@ -19,12 +19,12 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import kitchenpos.integration.helper.InitIntegrationTest;
 import kitchenpos.menu.application.dto.response.MenuGroupResponse;
 import kitchenpos.menu.application.dto.response.MenuResponse;
+import kitchenpos.menu.application.dto.response.ProductResponse;
 import kitchenpos.order.application.dto.request.OrderCreateRequest;
 import kitchenpos.order.application.dto.request.OrderLineItemRequest;
 import kitchenpos.order.application.dto.request.OrderStatusChangeRequest;
 import kitchenpos.order.application.dto.response.OrderResponse;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.product.application.dto.response.ProductResponse;
 import kitchenpos.table.application.dto.request.OrderTableChangeEmptyRequest;
 import kitchenpos.table.application.dto.request.OrderTableChangeNumberOfGuestRequest;
 import kitchenpos.table.application.dto.request.OrderTableCreateRequest;
