+++ src/test/java/kitchenpos/integration/fixture/ProductAPIFixture.java
@@ -5,8 +5,8 @@ import static io.restassured.RestAssured.given;
 import io.restassured.common.mapper.TypeRef;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import kitchenpos.product.application.dto.request.ProductCreateRequest;
-import kitchenpos.product.application.dto.response.ProductResponse;
+import kitchenpos.menu.application.dto.request.ProductCreateRequest;
+import kitchenpos.menu.application.dto.response.ProductResponse;
 import org.springframework.http.MediaType;
 import java.math.BigDecimal;
 import java.util.List;
